made a parent fragment for lineup
